feat(cli): remove listable dependency from CLI
