+++ packages/changed/tsconfig.json
@@ -9,5 +9,5 @@
   "filesGlob": ["./src/**/*.ts"],
   "include": ["src/**/*.ts"],
   "exclude": ["dist", "node_modules", "**/__mocks__/**", "**/__tests__/**"],
-  "references": [{ "path": "../core" }, { "path": "../list" }, { "path": "../listable" }]
+  "references": [{ "path": "../cli" }, { "path": "../core" }, { "path": "../list" }, { "path": "../listable" }]
 }
