+++ packages/cli/tsconfig.json
@@ -9,11 +9,5 @@
   "filesGlob": ["./src/**/*.ts"],
   "include": ["src/**/*.ts"],
   "exclude": ["dist", "node_modules", "**/__mocks__/**", "**/__tests__/**"],
-  "references": [
-    { "path": "../core" },
-    { "path": "../init" },
-    { "path": "../listable" },
-    { "path": "../publish" },
-    { "path": "../version" }
-  ]
+  "references": [{ "path": "../core" }, { "path": "../init" }]
 }
