+++ packages/listable/src/index.ts
@@ -1,8 +1,6 @@
 import { listableFormat } from './lib/listable-format.js';
-import { listableOptions } from './lib/listable-options.js';
 
 const listable = {
   format: listableFormat,
-  options: listableOptions,
 };
 export { listable };
